feat(Exhaustive Registry: frontend): add exhaustive exports list, move export list and modal logic to context components to be able to mutualise the components
