+++ front/src/dashboard/registry/ExportModal.tsx
@@ -1,34 +1,15 @@
-import React, { useEffect, useMemo, useState } from "react";
-import { z } from "zod";
-import { useMutation, useQuery } from "@apollo/client";
+import React from "react";
 import styles from "./MyExports.module.scss";
 import {
   DeclarationType,
   RegistryExportFormat,
-  Mutation,
-  MutationGenerateRegistryV2ExportArgs,
-  Query,
   RegistryV2ExportWasteType,
-  RegistryV2ExportType,
-  CompanyType
+  RegistryV2ExportType
 } from "@td/codegen-ui";
-import {
-  GENERATE_REGISTRY_V2_EXPORT,
-  GET_MY_COMPANIES,
-  GET_REGISTRY_V2_EXPORTS
-} from "./shared";
-import { FieldError, useForm } from "react-hook-form";
+
+import { FieldError } from "react-hook-form";
 import { datetimeToYYYYMMDD } from "../../Apps/Dashboard/Validation/BSPaoh/paohUtils";
-import { zodResolver } from "@hookform/resolvers/zod";
-import {
-  format,
-  getYear,
-  startOfYear,
-  endOfYear,
-  subYears,
-  endOfDay,
-  startOfDay
-} from "date-fns";
+import { format, getYear, startOfYear, endOfYear, subYears } from "date-fns";
 import { Modal } from "../../common/components";
 import Button from "@codegouvfr/react-dsfr/Button";
 import { InlineError } from "../../Apps/common/Components/Error/Error";
@@ -40,44 +21,11 @@ import Checkbox from "@codegouvfr/react-dsfr/Checkbox";
 import { WasteCodeSwitcher } from "./WasteCodeSwitcher";
 import { RegistryCompanySwitcher } from "./RegistryCompanySwitcher";
 import Alert from "@codegouvfr/react-dsfr/Alert";
-import { REGISTRY_DELEGATIONS } from "../../Apps/common/queries/registryDelegation/queries";
-
-type Props = { isOpen: boolean; onClose: () => void };
+import { getDeclarationTypeWording, getRegistryTypeWording } from "./MyExports";
+import { useRegistryExportModal } from "./RegistryV2ExportModalContext";
 
 const displayError = (error: FieldError | undefined) => {
   return error ? error.message : null;
-};
-
-const getRegistryTypeWording = (registryType: RegistryV2ExportType): string => {
-  switch (registryType) {
-    case RegistryV2ExportType.Ssd:
-      return `Sortie de statut de déchet`;
-    case RegistryV2ExportType.Incoming:
-      return `Registre entrant`;
-    case RegistryV2ExportType.Managed:
-      return `Registre géré`;
-    case RegistryV2ExportType.Outgoing:
-      return `Registre sortant`;
-    case RegistryV2ExportType.Transported:
-      return `Registre transporté`;
-    default:
-      return `Registre exhaustif`;
-  }
-};
-
-const getDeclarationTypeWording = (
-  declarationType: DeclarationType
-): string => {
-  switch (declarationType) {
-    case DeclarationType.All:
-      return `Tous`;
-    case DeclarationType.Bsd:
-      return `Tracé (bordereaux)`;
-    case DeclarationType.Registry:
-      return `Déclaré (registre national)`;
-    default:
-      return `Tous`;
-  }
 };
 
 const getFilterStateForRegistryType = (
@@ -99,48 +47,6 @@ const getFilterStateForRegistryType = (
   }
   return {
     disabled: false
-  };
-};
-
-const getDefaultsForRegistryType = (
-  registryType: RegistryV2ExportType
-): {
-  wasteTypes?: [RegistryV2ExportWasteType, ...RegistryV2ExportWasteType[]];
-  declarationType?: DeclarationType;
-} => {
-  if (registryType === RegistryV2ExportType.Ssd) {
-    return {
-      wasteTypes: [
-        RegistryV2ExportWasteType.Dnd,
-        RegistryV2ExportWasteType.Dd,
-        RegistryV2ExportWasteType.Texs
-      ],
-      declarationType: DeclarationType.Registry
-    };
-  } else if (registryType === RegistryV2ExportType.Incoming) {
-    return {
-      declarationType: DeclarationType.All
-    };
-  } else if (registryType === RegistryV2ExportType.Managed) {
-    return {
-      declarationType: DeclarationType.All
-    };
-  } else if (registryType === RegistryV2ExportType.Outgoing) {
-    return {
-      declarationType: DeclarationType.All
-    };
-  } else if (registryType === RegistryV2ExportType.Transported) {
-    return {
-      declarationType: DeclarationType.All
-    };
-  }
-  return {
-    wasteTypes: [
-      RegistryV2ExportWasteType.Dnd,
-      RegistryV2ExportWasteType.Dd,
-      RegistryV2ExportWasteType.Texs
-    ],
-    declarationType: DeclarationType.Registry
   };
 };
 
@@ -169,79 +75,9 @@ const getDateButtons = () => {
   ];
 };
 
-const getSchema = () =>
-  z
-    .object({
-      companyOrgId: z.string({ required_error: "Ce champ est requis" }),
-      isDelegation: z.boolean(),
-      delegateSiret: z.string().nullable(),
-      startDate: z.coerce
-        .date({
-          required_error: "La date de début est requise",
-          invalid_type_error: "La date de début est invalide"
-        })
-        .max(new Date(), {
-          message: "La date de début ne peut pas être dans le futur"
-        })
-        .transform(val => val.toISOString()),
-      // Date & "" hack: https://github.com/colinhacks/zod/issues/1721
-      endDate: z.preprocess(
-        arg => (arg === "" ? null : arg),
-        z.coerce
-          .date({
-            invalid_type_error: "La date de fin est invalide"
-          })
-          .max(new Date(), {
-            message: "La date de fin ne peut pas être dans le futur"
-          })
-          .transform(val => {
-            if (val) return val.toISOString();
-            return val;
-          })
-          .nullish()
-      ),
-      registryType: z.nativeEnum(RegistryV2ExportType),
-      format: z.nativeEnum(RegistryExportFormat),
-      declarationType: z.nativeEnum(DeclarationType),
-      wasteTypes: z.nativeEnum(RegistryV2ExportWasteType).array().nonempty({
-        message: "Veullez sélectionner au moins un type de déchet"
-      }),
-      wasteCodes: z.string().array()
-    })
-    .superRefine(({ startDate, endDate }, ctx) => {
-      if (startDate && endDate) {
-        if (new Date(startDate) > new Date(endDate)) {
-          ctx.addIssue({
-            code: z.ZodIssueCode.custom,
-            path: ["startDate"],
-            message: "La date de début doit être avant la date de fin."
-          });
-          ctx.addIssue({
-            code: z.ZodIssueCode.custom,
-            path: ["endDate"],
-            message: "La date de début doit être avant la date de fin."
-          });
-        }
-      }
-    });
-
-const getDefaultValues = () => ({
-  companyOrgId: "all",
-  isDelegation: false,
-  delegateSiret: null,
-  startDate: format(startOfYear(new Date()), "yyyy-MM-dd"),
-  registryType: RegistryV2ExportType.Incoming,
-  format: RegistryExportFormat.Csv,
-  declarationType: DeclarationType.All,
-  wasteTypes: [
-    RegistryV2ExportWasteType.Dd,
-    RegistryV2ExportWasteType.Dnd,
-    RegistryV2ExportWasteType.Texs
-  ],
-  wasteCodes: []
-});
-
-const getDateDescription = (registryType: RegistryV2ExportType): string => {
+const getDateDescription = (
+  registryType: RegistryV2ExportType | null
+): string => {
   switch (registryType) {
     case RegistryV2ExportType.Ssd:
       return "La date d'utilisation ou d'expédition est prise en compte.";
@@ -251,257 +87,48 @@ const getDateDescription = (registryType: RegistryV2ExportType): string => {
     case RegistryV2ExportType.Transported:
     case RegistryV2ExportType.Managed:
       return "La date d'expédition est prise en compte. ";
+    default:
+      return "La date de dernière mise à jour du bordereau est prise en compte.";
   }
-  return "";
 };
 
-export function ExportModal({ isOpen, onClose }: Props) {
-  // const [companies, setCompanies] = useState<ExportCompany[]>([]);
-  const [error, setError] = useState<string | null>(null);
-
-  const [generateExport, { loading: generateLoading }] = useMutation<
-    Pick<Mutation, "generateRegistryV2Export">,
-    Omit<MutationGenerateRegistryV2ExportArgs, "where"> & {
-      declarationType: DeclarationType;
-      wasteTypes: RegistryV2ExportWasteType[] | null;
-      wasteCodes: string[] | null;
-    }
-  >(GENERATE_REGISTRY_V2_EXPORT, {
-    refetchQueries: [GET_REGISTRY_V2_EXPORTS]
-  });
-
-  const validationSchema = getSchema();
+export function ExportModal() {
+  const {
+    type,
+    isOpen,
+    onClose,
+    isLoading,
+    companiesError,
+    error,
+    methods,
+    submit,
+    registryDelegationsData,
+    registryDelegationsLoading,
+    possibleExportTypes
+  } = useRegistryExportModal();
   const {
     register,
-    handleSubmit,
+    formState: { errors },
     setValue,
-    watch,
-    reset,
-    formState: { errors, isSubmitting }
-  } = useForm<z.infer<typeof validationSchema>>({
-    defaultValues: getDefaultValues(),
-    resolver: zodResolver(validationSchema)
-  });
-
-  const companyOrgId = watch("companyOrgId");
-  const isDelegation = watch("isDelegation");
-  const registryType = watch("registryType");
-  const startDate = watch("startDate");
-
-  const {
-    data: registryDelegationsData,
-    loading: registryDelegationsLoading,
-    error: registryDelegationsError
-  } = useQuery<Pick<Query, "registryDelegations">>(REGISTRY_DELEGATIONS, {
-    variables: {
-      where: {
-        delegatorOrgId: companyOrgId,
-        givenToMe: true,
-        activeOnly: true
-      }
-    },
-    skip: !isDelegation,
-    fetchPolicy: "network-only"
-  });
-
-  const { data: companiesData, error: companiesError } = useQuery<
-    Pick<Query, "myCompanies">
-  >(GET_MY_COMPANIES, {
-    variables: {
-      search: companyOrgId
-    },
-    skip: isDelegation,
-    fetchPolicy: "network-only"
-  });
-
-  const closeAndReset = () => {
-    setError(null);
-    reset(getDefaultValues());
-    onClose();
-  };
-
-  // get a list of possible export types for the selected company (or all companies) to grey out
-  // the export types that are impossible
-  const possibleExportTypes = useMemo(() => {
-    if (companyOrgId === "all") {
-      return [
-        RegistryV2ExportType.Outgoing,
-        RegistryV2ExportType.Incoming,
-        RegistryV2ExportType.Transported,
-        RegistryV2ExportType.Managed,
-        RegistryV2ExportType.Ssd
-      ];
-    }
-    const companies = isDelegation
-      ? registryDelegationsData?.registryDelegations.edges.map(
-          d => d.node.delegator
-        )
-      : companiesData?.myCompanies.edges.map(c => c.node);
-    const selectedCompany = companies?.find(c => c.orgId === companyOrgId);
-    if (!selectedCompany) {
-      return [
-        RegistryV2ExportType.Outgoing,
-        RegistryV2ExportType.Incoming,
-        RegistryV2ExportType.Transported,
-        RegistryV2ExportType.Managed,
-        RegistryV2ExportType.Ssd
-      ];
-    }
-    const companyTypes = selectedCompany.companyTypes;
-
-    const exportTypes: RegistryV2ExportType[] = [];
-
-    if (
-      companyTypes.filter(t =>
-        [
-          CompanyType.Producer,
-          CompanyType.WasteVehicles,
-          CompanyType.WasteCenter,
-          CompanyType.Collector,
-          CompanyType.Worker,
-          CompanyType.EcoOrganisme
-        ].includes(t)
-      ).length > 0
-    ) {
-      exportTypes.push(RegistryV2ExportType.Outgoing);
-    }
-
-    if (
-      companyTypes.filter(t =>
-        [
-          CompanyType.Wasteprocessor,
-          CompanyType.WasteVehicles,
-          CompanyType.Collector
-        ].includes(t)
-      ).length > 0
-    ) {
-      exportTypes.push(RegistryV2ExportType.Incoming);
-    }
-
-    if (companyTypes.includes(CompanyType.Transporter)) {
-      exportTypes.push(RegistryV2ExportType.Transported);
-    }
-
-    if (
-      companyTypes.includes(CompanyType.Trader) ||
-      companyTypes.includes(CompanyType.Broker) ||
-      companyTypes.includes(CompanyType.Intermediary)
-    ) {
-      exportTypes.push(RegistryV2ExportType.Managed);
-    }
-    exportTypes.push(RegistryV2ExportType.Ssd);
-    return exportTypes;
-  }, [companyOrgId, isDelegation, companiesData, registryDelegationsData]);
-
-  // if the registry type is not in the possible export types, set the first possible export type
-  useEffect(() => {
-    if (!possibleExportTypes.includes(registryType)) {
-      setValue("registryType", possibleExportTypes[0]);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [possibleExportTypes, registryType]);
-
-  // set the default values for waste types and declaration type depending on the registry type
-  useEffect(() => {
-    const defaults = getDefaultsForRegistryType(registryType);
-    if (defaults) {
-      if (defaults.wasteTypes) {
-        setValue("wasteTypes", defaults.wasteTypes);
-      }
-      if (defaults.declarationType) {
-        setValue("declarationType", defaults.declarationType);
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [registryType]);
-
-  const onSubmit = async (input: z.infer<typeof validationSchema>) => {
-    setError(null);
-    const {
-      companyOrgId,
-      isDelegation,
-      registryType,
-      format,
-      startDate,
-      endDate,
-      declarationType,
-      wasteTypes,
-      wasteCodes
-    } = input;
-    let delegateSiret = input.delegateSiret;
-    if (isDelegation) {
-      if (!delegateSiret) {
-        if (registryDelegationsData?.registryDelegations.edges.length === 1) {
-          const tmpSiret =
-            registryDelegationsData.registryDelegations.edges[0].node.delegate
-              .siret;
-          if (tmpSiret) {
-            delegateSiret = tmpSiret;
-          }
-        } else {
-          setError("Veuillez sélectionner un délégataire");
-          return;
-        }
-      }
-    }
-    const siret = companyOrgId === "all" ? null : companyOrgId;
-    // push the dates to the extremities of days so we include the days entered in the inputs
-    const startOfDayStartDate = startOfDay(new Date(startDate)).toISOString();
-    const endOfDayEndDate = endDate
-      ? endOfDay(new Date(endDate)).toISOString()
-      : null;
-
-    if (
-      registryType !== RegistryV2ExportType.Ssd &&
-      registryType !== RegistryV2ExportType.Incoming &&
-      registryType !== RegistryV2ExportType.Outgoing &&
-      registryType !== RegistryV2ExportType.Transported &&
-      registryType !== RegistryV2ExportType.Managed
-    ) {
-      setError(
-        "Seuls les exports SSD, entrants, sortants, transportés et gérés sont supportés pour le moment"
-      );
-      return;
-    }
-    await generateExport({
-      variables: {
-        siret,
-        delegateSiret,
-        registryType, // RegistryV2ExportType.Ssd
-        format, //RegistryExportFormat.Csv
-        dateRange: {
-          _gte: startOfDayStartDate,
-          _lt: endOfDayEndDate
-        },
-        declarationType, // DeclarationType.All
-        wasteTypes, //RegistryV2ExportWasteType[]
-        wasteCodes
-      },
-      onCompleted: () => {
-        closeAndReset();
-      },
-      onError: err => setError(err.message)
-    });
-  };
-
-  const isLoading = isSubmitting || generateLoading;
+    watch
+  } = methods;
   const dateButtons = getDateButtons();
-
+  const isDelegation = type === "registryV2" ? watch("isDelegation") : false;
+  const registryType = type === "registryV2" ? watch("registryType") : null;
+  const startDate = watch("startDate");
   return (
     <Modal
       title="Exporter"
       ariaLabel="Exporter un registre"
-      onClose={closeAndReset}
+      onClose={onClose}
       closeLabel="Annuler"
       isOpen={isOpen}
       size="M"
     >
-      {companiesError || registryDelegationsError ? (
-        <InlineError
-          apolloError={(companiesError || registryDelegationsError)!}
-        />
+      {companiesError ? (
+        <InlineError apolloError={companiesError} />
       ) : (
-        <form onSubmit={handleSubmit(onSubmit)}>
+        <form onSubmit={submit}>
           <div className="fr-mb-8v">
             <RegistryCompanySwitcher
               onCompanySelect={(orgId, isDelegation) => {
@@ -515,7 +142,7 @@ export function ExportModal({ isOpen, onClose }: Props) {
                 name: "Tous les établissements"
               }}
             />
-            {isDelegation ? (
+            {type === "registryV2" && isDelegation ? (
               registryDelegationsLoading ? (
                 <div className="fr-mt-2v">
                   <InlineLoader size={32} />
@@ -577,103 +204,112 @@ export function ExportModal({ isOpen, onClose }: Props) {
               ) : null
             ) : null}
           </div>
-          <div className="fr-mb-8v">
-            <Select
-              label="Type de registre"
-              disabled={isLoading}
-              nativeSelectProps={{
-                ...register("registryType")
-              }}
-            >
-              {[
-                RegistryV2ExportType.Incoming,
-                RegistryV2ExportType.Outgoing,
-                RegistryV2ExportType.Transported,
-                RegistryV2ExportType.Managed,
-                RegistryV2ExportType.Ssd
-              ].map(key => (
-                <option
-                  value={key}
-                  key={key}
-                  disabled={!possibleExportTypes.includes(key)}
+          {type === "registryV2" && (
+            <>
+              <div className="fr-mb-8v">
+                <Select
+                  label="Type de registre"
+                  disabled={isLoading}
+                  nativeSelectProps={{
+                    ...register("registryType")
+                  }}
                 >
-                  {getRegistryTypeWording(key)}
-                </option>
-              ))}
-            </Select>
-          </div>
-          <div className="fr-container--fluid fr-mb-8v">
-            <Select
-              label="Type de déclaration"
-              disabled={
-                isLoading ||
-                getFilterStateForRegistryType(registryType, "declarationType")
-                  .disabled
-              }
-              nativeSelectProps={{
-                ...register("declarationType")
-              }}
-            >
-              {Object.keys(DeclarationType).map(key => (
-                <option value={DeclarationType[key]} key={DeclarationType[key]}>
-                  {getDeclarationTypeWording(DeclarationType[key])}
-                </option>
-              ))}
-            </Select>
-          </div>
-          <div className="fr-container--fluid">
-            <Checkbox
-              hintText="Sélectionner au moins un type de déchets"
-              legend="Type de déchets"
-              disabled={isLoading}
-              options={[
-                {
-                  label: "Déchets non dangereux",
-                  nativeInputProps: {
-                    value: RegistryV2ExportWasteType.Dnd,
-                    disabled: getFilterStateForRegistryType(
-                      registryType,
-                      "wasteTypes.dnd"
-                    ).disabled,
-                    ...register("wasteTypes")
+                  {[
+                    RegistryV2ExportType.Incoming,
+                    RegistryV2ExportType.Outgoing,
+                    RegistryV2ExportType.Transported,
+                    RegistryV2ExportType.Managed,
+                    RegistryV2ExportType.Ssd
+                  ].map(key => (
+                    <option
+                      value={key}
+                      key={key}
+                      disabled={!possibleExportTypes.includes(key)}
+                    >
+                      {getRegistryTypeWording(key)}
+                    </option>
+                  ))}
+                </Select>
+              </div>
+              <div className="fr-container--fluid fr-mb-8v">
+                <Select
+                  label="Type de déclaration"
+                  disabled={
+                    isLoading ||
+                    getFilterStateForRegistryType(
+                      registryType as RegistryV2ExportType,
+                      "declarationType"
+                    ).disabled
                   }
-                },
-                {
-                  label: "Déchets dangereux",
-                  nativeInputProps: {
-                    value: RegistryV2ExportWasteType.Dd,
-                    disabled: getFilterStateForRegistryType(
-                      registryType,
-                      "wasteTypes.dd"
-                    ).disabled,
-                    ...register("wasteTypes")
-                  }
-                },
-                {
-                  label: "Terres et sédiments",
-                  nativeInputProps: {
-                    value: RegistryV2ExportWasteType.Texs,
-                    disabled: getFilterStateForRegistryType(
-                      registryType,
-                      "wasteTypes.texs"
-                    ).disabled,
-                    ...register("wasteTypes")
-                  }
-                }
-              ]}
-            />
-          </div>
-          <div className="fr-mb-8v">
-            <WasteCodeSwitcher
-              id={"wasteCodeSwitcher"}
-              onSelectChange={wasteCodes => {
-                setValue(
-                  "wasteCodes",
-                  wasteCodes.map(({ code }) => code)
-                );
-              }}
-            />
-          </div>
+                  nativeSelectProps={{
+                    ...register("declarationType")
+                  }}
+                >
+                  {Object.keys(DeclarationType).map(key => (
+                    <option
+                      value={DeclarationType[key]}
+                      key={DeclarationType[key]}
+                    >
+                      {getDeclarationTypeWording(DeclarationType[key])}
+                    </option>
+                  ))}
+                </Select>
+              </div>
+              <div className="fr-container--fluid">
+                <Checkbox
+                  hintText="Sélectionner au moins un type de déchets"
+                  legend="Type de déchets"
+                  disabled={isLoading}
+                  options={[
+                    {
+                      label: "Déchets non dangereux",
+                      nativeInputProps: {
+                        value: RegistryV2ExportWasteType.Dnd,
+                        disabled: getFilterStateForRegistryType(
+                          registryType as RegistryV2ExportType,
+                          "wasteTypes.dnd"
+                        ).disabled,
+                        ...register("wasteTypes")
+                      }
+                    },
+                    {
+                      label: "Déchets dangereux",
+                      nativeInputProps: {
+                        value: RegistryV2ExportWasteType.Dd,
+                        disabled: getFilterStateForRegistryType(
+                          registryType as RegistryV2ExportType,
+                          "wasteTypes.dd"
+                        ).disabled,
+                        ...register("wasteTypes")
+                      }
+                    },
+                    {
+                      label: "Terres et sédiments",
+                      nativeInputProps: {
+                        value: RegistryV2ExportWasteType.Texs,
+                        disabled: getFilterStateForRegistryType(
+                          registryType as RegistryV2ExportType,
+                          "wasteTypes.texs"
+                        ).disabled,
+                        ...register("wasteTypes")
+                      }
+                    }
+                  ]}
+                />
+              </div>
+              <div className="fr-mb-8v">
+                <WasteCodeSwitcher
+                  id={"wasteCodeSwitcher"}
+                  onSelectChange={wasteCodes => {
+                    setValue(
+                      "wasteCodes",
+                      wasteCodes.map(({ code }) => code)
+                    );
+                  }}
+                />
+              </div>
+            </>
+          )}
           <h6 className="fr-h6">{`Période concernée`}</h6>
           <div className={classNames(["fr-mb-8v", styles.dateButtons])}>
             {dateButtons.map((dateButton, key) => (
@@ -768,11 +404,7 @@ export function ExportModal({ isOpen, onClose }: Props) {
         </form>
       )}
       <div className="td-modal-actions">
-        <Button
-          priority="secondary"
-          disabled={isLoading}
-          onClick={closeAndReset}
-        >
+        <Button priority="secondary" disabled={isLoading} onClick={onClose}>
           Annuler
         </Button>
         <Button
@@ -780,7 +412,7 @@ export function ExportModal({ isOpen, onClose }: Props) {
           iconId="fr-icon-download-line"
           iconPosition="right"
           disabled={isLoading}
-          onClick={handleSubmit(onSubmit)}
+          onClick={submit}
         >
           Exporter
         </Button>
