+++ front/src/dashboard/registry/MyExports.tsx
@@ -1,33 +1,29 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { useLazyQuery, useQuery } from "@apollo/client";
+import React, { useMemo } from "react";
 import { ExportModal } from "./ExportModal";
 
 import styles from "./MyExports.module.scss";
 import {
   DeclarationType,
-  Query,
-  QueryRegistryV2ExportDownloadSignedUrlArgs,
   RegistryExportStatus,
-  RegistryV2ExportType
+  RegistryV2ExportType,
+  RegistryV2Export,
+  RegistryExhaustiveExport
 } from "@td/codegen-ui";
-import {
-  badges,
-  downloadFromSignedUrl,
-  GET_REGISTRY_V2_EXPORTS,
-  REGISTRY_V2_EXPORT_DOWNLOAD_SIGNED_URL
-} from "./shared";
+import { badges } from "./shared";
 
-import { format, getYear, startOfYear, endOfYear, subHours } from "date-fns";
+import { format, getYear, startOfYear, endOfYear } from "date-fns";
 import Button from "@codegouvfr/react-dsfr/Button";
 import Tooltip from "@codegouvfr/react-dsfr/Tooltip";
 import { InlineLoader } from "../../Apps/common/Components/Loader/Loaders";
 import Pagination from "@codegouvfr/react-dsfr/Pagination";
 import Alert from "@codegouvfr/react-dsfr/Alert";
 import RegistryTable from "./RegistryTable";
-import { useMedia } from "../../common/use-media";
-import { MEDIA_QUERIES } from "../../common/config";
+import { useRegistryExport } from "./RegistryV2ExportContext";
+import { useRegistryExportModal } from "./RegistryV2ExportModalContext";
 
-const getRegistryTypeWording = (registryType: RegistryV2ExportType): string => {
+export const getRegistryTypeWording = (
+  registryType: RegistryV2ExportType
+): string => {
   switch (registryType) {
     case RegistryV2ExportType.Ssd:
       return `Sortie de statut de déchet`;
@@ -39,12 +35,10 @@ const getRegistryTypeWording = (registryType: RegistryV2ExportType): string => {
       return `Registre sortant`;
     case RegistryV2ExportType.Transported:
       return `Registre transporté`;
-    default:
-      return `Registre exhaustif`;
   }
 };
 
-const getDeclarationTypeWording = (
+export const getDeclarationTypeWording = (
   declarationType: DeclarationType
 ): string => {
   switch (declarationType) {
@@ -91,167 +85,133 @@ const formatRegistryDates = (
   )}`;
 };
 
-const PAGE_SIZE = 20;
-
 export function MyExports() {
-  const [pageIndex, setPageIndex] = useState(0);
-  const [isExportModalOpen, setIsExportModalOpen] = useState(false);
-  const [downloadLoadingExportId, setDownloadLoadingExportId] = useState<
-    string | null
-  >(null);
   const {
-    data: exportsData,
-    loading: exportsLoading,
-    error,
+    type,
+    pageIndex,
+    pageCount,
+    downloadLoadingExportId,
+    exportsLoading,
+    registryExports,
+    downloadRegistryExportFile,
+    gotoPage,
     refetch,
-    startPolling,
-    stopPolling
-  } = useQuery<Pick<Query, "registryV2Exports">>(GET_REGISTRY_V2_EXPORTS, {
-    variables: { first: PAGE_SIZE },
-    fetchPolicy: "cache-and-network"
-  });
-  const registryExports = exportsData?.registryV2Exports?.edges;
-  const totalCount = exportsData?.registryV2Exports?.totalCount;
-  const pageCount = totalCount ? Math.ceil(totalCount / PAGE_SIZE) : 0;
+    error
+  } = useRegistryExport();
+  const { onOpen: onOpenExportModal } = useRegistryExportModal();
 
-  const [getDownloadLink] = useLazyQuery<
-    Pick<Query, "registryV2ExportDownloadSignedUrl">,
-    Partial<QueryRegistryV2ExportDownloadSignedUrlArgs>
-  >(REGISTRY_V2_EXPORT_DOWNLOAD_SIGNED_URL, { fetchPolicy: "no-cache" });
-
-  const downloadRegistryExportFile = useCallback(
-    async (exportId: string) => {
-      setDownloadLoadingExportId(exportId);
-      try {
-        const link = await getDownloadLink({
-          variables: { exportId }
-        });
-        downloadFromSignedUrl(
-          link.data?.registryV2ExportDownloadSignedUrl.signedUrl
-        );
-      } finally {
-        setDownloadLoadingExportId(null);
-      }
-    },
-    [setDownloadLoadingExportId, getDownloadLink]
-  );
-
-  const gotoPage = useCallback(
-    (page: number) => {
-      setPageIndex(page);
-      refetch({
-        skip: page * PAGE_SIZE
-      });
-    },
-    [setPageIndex, refetch]
-  );
-
-  useEffect(() => {
-    if (
-      registryExports?.some(
-        registryExport =>
-          (registryExport.node.status === RegistryExportStatus.Pending ||
-            registryExport.node.status === RegistryExportStatus.Started) &&
-          // condition to avoid infinite refetches on old exports that somehow never finished
-          new Date(registryExport.node.createdAt) > subHours(new Date(), 1)
-      )
-    ) {
-      startPolling(5000);
-    } else {
-      stopPolling();
-    }
-    return () => {
-      stopPolling();
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [registryExports]);
   const tableData = useMemo(() => {
-    return registryExports
-      ? registryExports.map(registryExport => [
-          <div>
-            <div>
-              {format(
-                new Date(registryExport.node.createdAt),
-                "dd/MM/yyyy HH:mm"
-              )}
-            </div>
-            <div className={styles.badge}>
-              {badges[registryExport.node.status]("export")}
-            </div>
-          </div>,
-          <div>
-            {(registryExport.node.status === RegistryExportStatus.Started ||
-              registryExport.node.status === RegistryExportStatus.Pending) &&
-            registryExport.node.companies.length > 1 ? null : (
-              <>
-                {[
-                  `${
-                    registryExport.node.companies[0]?.givenName &&
-                    registryExport.node.companies[0]?.givenName !== ""
-                      ? registryExport.node.companies[0]?.givenName
-                      : registryExport.node.companies[0]?.name
-                  } - ${registryExport.node.companies[0]?.orgId}`,
-                  ...(registryExport.node.companies.length > 1
-                    ? [
-                        `et ${registryExport.node.companies.length - 1} autre${
-                          registryExport.node.companies.length > 2 ? "s" : ""
-                        } `
-                      ]
-                    : [])
-                ].join(", ")}
-                {registryExport.node.companies.length > 1 ? (
-                  <Tooltip
-                    kind="hover"
-                    className={styles.prewrap}
-                    title={registryExport.node.companies
-                      .slice(1)
-                      .map(
-                        company =>
-                          `${
-                            company.givenName && company.givenName !== ""
-                              ? company.givenName
-                              : company.name
-                          } - ${company.orgId}`
-                      )
-                      .join(",\n")}
-                  />
-                ) : null}
-              </>
-            )}
-          </div>,
-          getRegistryTypeWording(registryExport.node.registryType),
-          getDeclarationTypeWording(registryExport.node.declarationType),
-          formatRegistryDates(
-            registryExport.node.createdAt,
-            registryExport.node.startDate,
-            registryExport.node.endDate
-          ),
-          registryExport.node.status === RegistryExportStatus.Pending ||
-          registryExport.node.status === RegistryExportStatus.Started ||
-          downloadLoadingExportId === registryExport.node.id ? (
-            <div className="tw-px-2" style={{ width: "fit-content" }}>
-              <InlineLoader size={32} />
-            </div>
-          ) : registryExport.node.status === RegistryExportStatus.Successful ? (
-            <div className="tw-px-2">
-              <Button
-                title="Télécharger"
-                priority="secondary"
-                iconId="fr-icon-download-line"
-                onClick={() =>
-                  downloadRegistryExportFile(registryExport.node.id)
-                }
-                size="small"
-              />
-            </div>
-          ) : (
-            ""
-          )
-        ])
-      : [];
-  }, [registryExports, downloadLoadingExportId, downloadRegistryExportFile]);
+    if (!registryExports) return [];
 
-  const isMobile = useMedia(`(max-width: ${MEDIA_QUERIES.handHeld})`);
+    const getCommonColumns = (
+      registryExport: RegistryV2Export | RegistryExhaustiveExport
+    ) => ({
+      date: (
+        <div>
+          <div>
+            {format(new Date(registryExport.createdAt), "dd/MM/yyyy HH:mm")}
+          </div>
+          <div className={styles.badge}>
+            {badges[registryExport.status]("export")}
+          </div>
+        </div>
+      ),
+      companies: (
+        <div>
+          {(registryExport.status === RegistryExportStatus.Started ||
+            registryExport.status === RegistryExportStatus.Pending) &&
+          registryExport.companies.length > 1 ? null : (
+            <>
+              {[
+                `${
+                  registryExport.companies[0]?.givenName &&
+                  registryExport.companies[0]?.givenName !== ""
+                    ? registryExport.companies[0]?.givenName
+                    : registryExport.companies[0]?.name
+                } - ${registryExport.companies[0]?.orgId}`,
+                ...(registryExport.companies.length > 1
+                  ? [
+                      `et ${registryExport.companies.length - 1} autre${
+                        registryExport.companies.length > 2 ? "s" : ""
+                      } `
+                    ]
+                  : [])
+              ].join(", ")}
+              {registryExport.companies.length > 1 ? (
+                <Tooltip
+                  kind="hover"
+                  className={styles.prewrap}
+                  title={registryExport.companies
+                    .slice(1)
+                    .map(
+                      company =>
+                        `${
+                          company.givenName && company.givenName !== ""
+                            ? company.givenName
+                            : company.name
+                        } - ${company.orgId}`
+                    )
+                    .join(",\n")}
+                />
+              ) : null}
+            </>
+          )}
+        </div>
+      ),
+      period: formatRegistryDates(
+        registryExport.createdAt,
+        registryExport.startDate,
+        registryExport.endDate
+      ),
+      file:
+        registryExport.status === RegistryExportStatus.Pending ||
+        registryExport.status === RegistryExportStatus.Started ||
+        downloadLoadingExportId === registryExport.id ? (
+          <div className="tw-px-2" style={{ width: "fit-content" }}>
+            <InlineLoader size={32} />
+          </div>
+        ) : registryExport.status === RegistryExportStatus.Successful ? (
+          <div className="tw-px-2">
+            <Button
+              title="Télécharger"
+              priority="secondary"
+              iconId="fr-icon-download-line"
+              onClick={() => downloadRegistryExportFile(registryExport.id)}
+              size="small"
+            />
+          </div>
+        ) : (
+          ""
+        )
+    });
+
+    return registryExports.map(registryExport => {
+      const commonColumns = getCommonColumns(registryExport);
+
+      if (type === "registryV2") {
+        return [
+          commonColumns.date,
+          commonColumns.companies,
+          getRegistryTypeWording(registryExport.registryType),
+          getDeclarationTypeWording(registryExport.declarationType),
+          commonColumns.period,
+          commonColumns.file
+        ];
+      } else {
+        return [
+          commonColumns.date,
+          commonColumns.companies,
+          commonColumns.period,
+          commonColumns.file
+        ];
+      }
+    });
+  }, [
+    registryExports,
+    downloadLoadingExportId,
+    downloadRegistryExportFile,
+    type
+  ]);
 
   return (
     <>
@@ -264,7 +224,7 @@ export function MyExports() {
                 priority="primary"
                 iconId="fr-icon-download-line"
                 iconPosition="right"
-                onClick={() => setIsExportModalOpen(true)}
+                onClick={onOpenExportModal}
               >
                 Exporter
               </Button>
@@ -310,36 +270,18 @@ export function MyExports() {
               headers={[
                 "Date",
                 "Établissements",
-                "Type de registre",
-                "Type de déclaration",
+                ...(type === "registryV2"
+                  ? ["Type de registre", "Type de déclaration"]
+                  : []),
                 "Période",
                 "Fichier"
               ]}
-              fixed={!isMobile}
             />
           ) : (
             <InlineLoader />
           )}
         </div>
 
-        {/* <div className="tw-p-6">
-          {!exportsLoading ? (
-            <Table
-              bordered
-              caption="Exports récents"
-              className={styles.fullWidthTable}
-              data={tableData}
-              headers={[
-                "Date",
-                "Établissements",
-                "Type de registre",
-                "Type de déclaration",
-                "Période",
-                "Fichier"
-              ]}
-            />
-          ) : <InlineLoader />}
-        </div> */}
         <div className="tw-flex tw-justify-center">
           <Pagination
             showFirstLast
@@ -357,13 +299,7 @@ export function MyExports() {
           />
         </div>
       </>
-      <ExportModal
-        isOpen={isExportModalOpen}
-        onClose={() => {
-          setIsExportModalOpen(false);
-          refetch();
-        }}
-      />
+      <ExportModal />
     </>
   );
 }
