+++ front/src/dashboard/registry/RegistryRoutes.tsx
@@ -16,9 +16,11 @@ import { MyExports } from "./MyExports";
 import { MyImports } from "./MyImports";
 import { MyLines } from "./myLines/MyLines";
 import { FormContainer } from "./myLines/FormContainer";
-import Exports from "../exports/Registry";
 import "../../Apps/Dashboard/dashboard.scss";
 import { RegistryImportType } from "@td/codegen-ui";
+import { RegistryV2ExportProvider } from "./RegistryV2ExportContext";
+import { RegistryExhaustiveExportProvider } from "./RegistryExhaustiveExportContext";
+import { RegistryV2ExportModalProvider } from "./RegistryV2ExportModalContext";
 
 const toRelative = route => {
   return getRelativeRoute(routes.registry_new.index, route);
@@ -75,7 +77,13 @@ export default function RegistryRoutes() {
 
           <Route
             path={toRelative(routes.registry_new.export)}
-            element={<MyExports />}
+            element={
+              <RegistryV2ExportProvider>
+                <RegistryV2ExportModalProvider>
+                  <MyExports />
+                </RegistryV2ExportModalProvider>
+              </RegistryV2ExportProvider>
+            }
           />
 
           <Route
@@ -85,7 +93,13 @@ export default function RegistryRoutes() {
 
           <Route
             path={toRelative(routes.registry_new.exhaustive)}
-            element={<Exports />}
+            element={
+              <RegistryExhaustiveExportProvider>
+                <RegistryV2ExportModalProvider>
+                  <MyExports />
+                </RegistryV2ExportModalProvider>
+              </RegistryExhaustiveExportProvider>
+            }
           />
 
           <Route
